--- components/intro.tsx
@@ -1,16 +1,0 @@
-const Intro = () => {
-  return (
-    
-    <section className="flex-col md:flex-row flex items-center md:justify-between mt-16 mb-16 md:mb-12">
-       {/* <h1 className="text-5xl md:text-8xl font-bold tracking-tighter leading-tight md:pr-8">
-        Avgangsutstilling 2023
-      </h1>
-      <h4 className="text-center md:text-left text-lg mt-5 md:pl-8">
-        Tekst om utstillingen
-      </h4> */}
-    </section>
-    
-  )
-}
-
-export default Intro
