+++ pages/index.tsx
@@ -2,7 +2,7 @@ import Container from "../components/layout/container"
 import Layout from "../components/layout/layout"
 import Head from "next/head"
 import HomeContent from "../components/studyprograminfo"
-import BottomMap from "../components/homemap"
+import BottomMap from "../components/BottomMap"
 import HeroDel from "../components/temporary/hero-del"
 import StudyProgrammeInfo from "../components/StudyProgrammeInfo/StudyProgrammeInfo"
